refactor: extract single username gradebook fetch
