+++ lms/djangoapps/grades/rest_api/v1/gradebook_views.py
@@ -19,6 +19,8 @@ from rest_framework import status
 from rest_framework.generics import GenericAPIView
 from rest_framework.response import Response
 from rest_framework.views import APIView
+from xmodule.modulestore.django import modulestore
+from xmodule.util.misc import get_default_short_labeler
 
 from common.djangoapps.student.auth import has_course_author_access
 from common.djangoapps.student.models import CourseEnrollment
@@ -39,7 +41,12 @@ from lms.djangoapps.grades.api import events as grades_events
 from lms.djangoapps.grades.api import gradebook_can_see_bulk_management as can_see_bulk_management
 from lms.djangoapps.grades.api import is_writable_gradebook_enabled, prefetch_course_and_subsection_grades
 from lms.djangoapps.grades.course_data import CourseData
-from lms.djangoapps.grades.grade_utils import are_grades_frozen
+from lms.djangoapps.grades.grade_utils import (
+    are_grades_frozen,
+    bulk_gradebook_view_context,
+    gradebook_entry,
+    paginate_users
+)
 # TODO these imports break abstraction of the core Grades layer. This code needs
 # to be refactored so Gradebook views only access public Grades APIs.
 from lms.djangoapps.grades.models import (
@@ -66,27 +73,8 @@ from openedx.core.lib.api.view_utils import (
     view_auth_classes
 )
 from openedx.core.lib.cache_utils import request_cached
-from xmodule.modulestore.django import modulestore
-from xmodule.util.misc import get_default_short_labeler
 
 log = logging.getLogger(__name__)
-
-
-@contextmanager
-def bulk_gradebook_view_context(course_key, users):
-    """
-    Prefetches all course and subsection grades in the given course for the given
-    list of users, also, fetch all the score relavant data,
-    storing the result in a RequestCache and deleting grades on context exit.
-    """
-    prefetch_course_and_subsection_grades(course_key, users)
-    CourseEnrollment.bulk_fetch_enrollment_states(users, course_key)
-    cohorts.bulk_cache_cohorts(course_key, users)
-    BulkRoleCache.prefetch(users)
-    try:
-        yield
-    finally:
-        clear_prefetched_course_and_subsection_grades(course_key)
 
 
 def verify_writable_gradebook_enabled(view_func):
@@ -219,6 +207,7 @@ def course_author_access_required(view):
         def my_view(request, course_key):
             # Some functionality ...
     """
+
     def _wrapper_view(self, request, course_id, *args, **kwargs):
         """
         Checks for course author access for the given course by the requesting user.
@@ -426,85 +415,6 @@ class GradebookView(GradeViewMixin, PaginatedAPIView):
 
     pagination_class = CourseEnrollmentPagination
 
-    def _section_breakdown(self, course, graded_subsections, course_grade):
-        """
-        Given a course_grade and a list of graded subsections for a given course,
-        returns a list of grade data broken down by subsection.
-
-        Args:
-            course: A Course Descriptor object
-            graded_subsections: A list of graded subsection objects in the given course.
-            course_grade: A CourseGrade object.
-        """
-        breakdown = []
-        default_labeler = get_default_short_labeler(course)
-
-        for subsection in graded_subsections:
-            subsection_grade = course_grade.subsection_grade(subsection.location)
-            short_label = default_labeler(subsection_grade.format)
-
-            attempted = False
-            score_earned = 0
-            score_possible = 0
-
-            # For ZeroSubsectionGrades, we don't want to crawl the subsection's
-            # subtree to find the problem scores specific to this user
-            # (ZeroSubsectionGrade.attempted_graded is always False).
-            # We've already fetched the whole course structure in a non-user-specific way
-            # when creating `graded_subsections`.  Looking at the problem scores
-            # specific to this user (the user in `course_grade.user`) would require
-            # us to re-fetch the user-specific course structure from the modulestore,
-            # which is a costly operation.  So we only drill into the `graded_total`
-            # attribute if the user has attempted this graded subsection, or if there
-            # has been a grade override applied.
-            if subsection_grade.attempted_graded or subsection_grade.override:
-                attempted = True
-                score_earned = subsection_grade.graded_total.earned
-                score_possible = subsection_grade.graded_total.possible
-
-            # TODO: https://openedx.atlassian.net/browse/EDUCATOR-3559 -- Some fields should be renamed, others removed:
-            # 'displayed_value' should maybe be 'description_percent'
-            # 'grade_description' should be 'description_ratio'
-            breakdown.append({
-                'attempted': attempted,
-                'category': subsection_grade.format,
-                'label': short_label,
-                'module_id': str(subsection_grade.location),
-                'percent': subsection_grade.percent_graded,
-                'score_earned': score_earned,
-                'score_possible': score_possible,
-                'subsection_name': subsection_grade.display_name,
-            })
-        return breakdown
-
-    def _gradebook_entry(self, user, course, graded_subsections, course_grade):
-        """
-        Returns a dictionary of course- and subsection-level grade data for
-        a given user in a given course.
-
-        Args:
-            user: A User object.
-            course: A Course Descriptor object.
-            graded_subsections: A list of graded subsections in the given course.
-            course_grade: A CourseGrade object.
-        """
-        user_entry = self._serialize_user_grade(user, course.id, course_grade)
-        breakdown = self._section_breakdown(course, graded_subsections, course_grade)
-
-        user_entry['section_breakdown'] = breakdown
-        user_entry['progress_page_url'] = reverse(
-            'student_progress',
-            kwargs=dict(course_id=str(course.id), student_id=user.id)
-        )
-        user_entry['user_id'] = user.id
-        user_entry['full_name'] = user.profile.name
-
-        external_user_key = get_external_key_by_user_and_course(user, course.id)
-        if external_user_key:
-            user_entry['external_user_key'] = external_user_key
-
-        return user_entry
-
     @verify_course_exists
     @verify_writable_gradebook_enabled
     @course_author_access_required
@@ -533,7 +443,7 @@ class GradebookView(GradeViewMixin, PaginatedAPIView):
                     course,
                     collected_block_structure=course_data.collected_structure
                 )
-            entry = self._gradebook_entry(grade_user, course, graded_subsections, course_grade)
+            entry = gradebook_entry(grade_user, course, graded_subsections, course_grade)
             serializer = StudentGradebookEntrySerializer(entry)
             return Response(serializer.data)
         else:
@@ -621,7 +531,7 @@ class GradebookView(GradeViewMixin, PaginatedAPIView):
 
             entries = []
             related_models = ['user']
-            users = self._paginate_users(course_key, q_objects, related_models, annotations=annotations)
+            users = paginate_users(course_key, q_objects, related_models, annotations=annotations)
 
             users_counts = self._get_users_counts(course_key, q_objects, annotations=annotations)
 
@@ -630,7 +540,7 @@ class GradebookView(GradeViewMixin, PaginatedAPIView):
                     users, course_key=course_key, collected_block_structure=course_data.collected_structure
                 ):
                     if not exc:
-                        entry = self._gradebook_entry(user, course, graded_subsections, course_grade)
+                        entry = gradebook_entry(user, course, graded_subsections, course_grade)
                         entries.append(entry)
 
             serializer = StudentGradebookEntrySerializer(entries, many=True)
