+++ lms/djangoapps/grades/rest_api/v1/utils.py
@@ -15,6 +15,7 @@ from rest_framework.response import Response
 from common.djangoapps.student.models import CourseEnrollment
 from common.djangoapps.util.query import use_read_replica_if_available
 from lms.djangoapps.grades.course_grade_factory import CourseGradeFactory
+from lms.djangoapps.grades.grade_utils import serialize_user_grade
 from openedx.core.lib.api.view_utils import DeveloperErrorViewMixin
 
 USER_MODEL = get_user_model()
@@ -58,6 +59,7 @@ class GradeViewMixin(DeveloperErrorViewMixin):
     """
     Mixin class for Grades related views.
     """
+
     def _get_single_user(self, request, course_key, user_id=None):
         """
         Returns a single USER_MODEL object corresponding to either the user_id provided, or if no id is provided,
@@ -127,55 +129,7 @@ class GradeViewMixin(DeveloperErrorViewMixin):
             A serializable list of grade responses
         """
         course_grade = CourseGradeFactory().read(grade_user, course_key=course_key)
-        return Response([self._serialize_user_grade(grade_user, course_key, course_grade)])
-
-    def _paginate_users(self, course_key, course_enrollment_filter=None, related_models=None, annotations=None):
-        """
-        Args:
-            course_key (CourseLocator): The course to retrieve grades for.
-            course_enrollment_filter: Optional list of Q objects to pass
-            to `CourseEnrollment.filter()`.
-            related_models: Optional list of related models to join to the CourseEnrollment table.
-            annotations: Optional dict of fields to add to the queryset via annotation
-
-        Returns:
-            A list of users, pulled from a paginated queryset of enrollments, who are enrolled in the given course.
-        """
-        queryset = CourseEnrollment.objects
-        if annotations:
-            queryset = queryset.annotate(**annotations)
-
-        filter_args = [
-            Q(course_id=course_key) & Q(is_active=True)
-        ]
-        filter_args.extend(course_enrollment_filter or [])
-
-        enrollments_in_course = use_read_replica_if_available(
-            queryset.filter(*filter_args)
-        )
-        if related_models:
-            enrollments_in_course = enrollments_in_course.select_related(*related_models)
-
-        paged_enrollments = self.paginate_queryset(enrollments_in_course)
-        retlist = []
-        for enrollment in paged_enrollments:
-            enrollment.user.enrollment_mode = enrollment.mode
-            retlist.append(enrollment.user)
-        return retlist
-
-    def _serialize_user_grade(self, user, course_key, course_grade):
-        """
-        Serialize a single grade to dict to use in Responses
-        """
-        return {
-            'username': user.username,
-            # per business requirements, email should only be visible for students in masters track only
-            'email': user.email if getattr(user, 'enrollment_mode', '') == 'masters' else '',
-            'course_id': str(course_key),
-            'passed': course_grade.passed,
-            'percent': course_grade.percent,
-            'letter_grade': course_grade.letter_grade,
-        }
+        return Response([serialize_user_grade(grade_user, course_key, course_grade)])
 
     def perform_authentication(self, request):
         """
